requirements.txt
fastapi
uvicorn
jinja2
python-multipart

nltk

google-generativeai

requests
pillow

python-dotenv
